requirements.txt
spotipy==2.25.1
python-dotenv==1.1.0
requests==2.32.3
pylint==2.17.5
pytest==7.4.0
pytest-cov==4.1.0
pytest-mock==3.12.0
pandas>=2.0.3
